Update solution for matrix-transpose

+++ matrix-transpose/matrix-transpose.py
@@ -6,7 +6,7 @@ def matrix_transpose(A):
     """
     # Write code here
     M, N = len(A), len(A[0])
-    x = np.zeros((N,M), dtype=np.int_)
+    x = np.zeros((N,M))
     for i in range(M):
         for j in range(N): x[j][i] = A[i][j]
     return x
